added delivery date and category to the backend

+++ backend/src/main/java/com/example/inventory/repository/InventoryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
+public interface InventoryRepository extends JpaRepository<Inventory, Long> {
 
     List<Inventory> findAllByIsDeletedFalse();
